+++ examples/history.ts
@@ -1,12 +1,21 @@
 import { createWriteStream } from 'fs';
 import { type SparkClient } from '@cspark/sdk';
 
+function find(spark: SparkClient) {
+  spark.service.log
+    .find('my-folder/my-service')
+    .then((response) => {
+      console.log(JSON.stringify(response.data, null, 2));
+    })
+    .catch(console.error);
+}
+
 function rehydrate(spark: SparkClient) {
   spark.service.log
     .rehydrate('my-folder/my-service', 'uuid')
     .then((response) => {
       console.log(JSON.stringify(response.data, null, 2));
-      const file = createWriteStream('my-rehydrated-subservices.xlsx');
+      const file = createWriteStream('my-rehydrated-services.xlsx');
       response.buffer.pipe(file);
     })
     .catch(console.error);
@@ -28,4 +37,4 @@ function download(spark: SparkClient) {
     .catch(console.error);
 }
 
-export default { rehydrate, download };
+export default { find, rehydrate, download };