+++ examples/usecases/1-1k-sync-apiv3-for-loop-json.ts
@@ -41,6 +41,7 @@ async function main() {
   const writer = createWriteStream(join(basePath, new Date().toISOString() + '_results.json'));
   const logger = new Logger('sync-batch');
   const spark = new Spark({ ...sparkOptions, logger: { logger } });
+  const service = spark.service;
 
   // This helps keep track of the request ID for each API call (not required).
   spark.config.interceptors.add({
@@ -61,7 +62,7 @@ async function main() {
   for (let i = 0; i < dataset.length; i++) {
     const inputs = dataset[i];
     try {
-      const response = await spark.service.execute<Inputs, Outputs>(serviceUri, { inputs });
+      const response = await service.execute<Inputs, Outputs>(serviceUri, { inputs });
       const result = { inputs, result: response.data };
 
       writer.write(JSON.stringify(result, null, 2) + (i < total - 1 ? ',\n' : ''));