+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version)
 for commit guidelines.
 
+## 0.1.2 (2024-05-15)
+
+- Fix get status of batch pipeline
+- Add get batch pipeline info method
+- Mask sensitive information (e.g., API key, OAuth credentials)
+
 ## 0.1.1 (2024-05-01)
 
 - Add support for async Batch API (experimental feature)